+++ mono/test/commands/ungroup_command_test.dart
@@ -9,6 +9,12 @@ import '../util/fakes.dart';
 import '../util/fs_fixtures.dart';
 
 void main() {
+  late WorkspaceConfig workspaceConfig;
+
+  setUp(() async {
+    workspaceConfig = const FileWorkspaceConfig();
+  });
+
   group('UngroupCommand', () {
     test('usage error when name missing', () async {
       final ws = await createTempWorkspace('mono_ungroup_');
@@ -23,6 +29,7 @@ void main() {
           out: outCap.sink,
           err: errCap.sink,
           prompter: FakePrompter(),
+          workspaceConfig: workspaceConfig,
         );
         expect(code, 2);
         expect(errCap.text, contains('Usage: mono ungroup <group_name>'));
@@ -47,6 +54,7 @@ void main() {
           out: outCap.sink,
           err: errCap.sink,
           prompter: FakePrompter(),
+          workspaceConfig: workspaceConfig,
         );
         expect(code, 2);
         expect(errCap.text, contains('Group "dev" does not exist.'));
@@ -73,6 +81,7 @@ void main() {
           out: outCap.sink,
           err: errCap.sink,
           prompter: FakePrompter(nextConfirm: false),
+          workspaceConfig: workspaceConfig,
         );
         expect(code, 1);
         expect(errCap.text, contains('Aborted.'));
@@ -101,6 +110,7 @@ void main() {
           out: outCap.sink,
           err: errCap.sink,
           prompter: FakePrompter(nextConfirm: true),
+          workspaceConfig: workspaceConfig,
         );
         expect(code, 0);
         expect(outCap.text, contains('Group "dev" removed.'));