+++ mono_core/lib/src/src.dart
@@ -18,6 +18,7 @@ export 'ports/system.dart';
 export 'ports/target_selector.dart';
 export 'ports/task_plugin.dart';
 export 'ports/version.dart';
+export 'ports/workspace_config.dart';
 export 'selector/target_expr.dart';
 export 'types/errors.dart';
 export 'types/types.dart';